sched: ems: fix return type of task_util.
The variable type of util_avg is unsigned long. Fix return type
of task util to avoid data loss.

+++ kernel/sched/ems/core.c
@@ -13,7 +13,7 @@
 #include "ems.h"
 #include "../sched.h"
 
-int task_util(struct task_struct *p)
+unsigned long task_util(struct task_struct *p)
 {
 	if (rt_task(p))
 		return p->rt.avg.util_avg;
